+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [Unreleased]
 
+### Removed
+- Remove the deprecated endpoints from version 0.9.15
+  [#1200](https://github.com/nextcloud/cookbook/pull/1200) @christianlupus
 
 
 ## 0.9.15 - 2022-09-08